requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
numpy==1.26.4

faiss-cpu==1.7.4
transformers==4.41.1
sentence-transformers==3.0.1
torch==2.2.1

reportlab==4.0.0
pydantic==1.10.13
python-multipart==0.0.9

TTS==0.13.0
openai-whisper==2023.12.1
